drm/amd/dal: Fix clock source remapping issue when DRM swaps crtcs
When disabling and enabling monitors in a specific order, eventually we'll get
to a point where the clock sources will be swapped.

BIOS appears to have an optimization where the updated phy on enable_output
doesn't take affect if it hasn't been disabled.

To fix this, reset the pipe if the clock sources change.

Signed-off-by: Jordan Lazare <Jordan.Lazare@amd.com>
Acked-by: Harry Wentland <harry.wentland@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>