drm/amd/dal: Switch in order setting scaler enable and number of taps.
Number of taps is not cleared when switching from scaled to non scaled mode.
THis could potentially lead to teprorary corruption when enbaling scaler next
time and still not setting the proper number of taps. Until scaler lock is used
as a work around reverse order of setting taps and enable scaler.

Signed-off-by: Andrey Grodzovsky <Andrey.Grodzovsky@amd.com>
Acked-by: Harry Wentland <harry.wentland@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>