Revert "UBUNTU: SAUCE: (efi-lockdown) efi: Add sysctls for secureboot and MokSBState"
This patch is no longer necessary, as the userspace tools have
been reworked to support getting this information from EFI
variables. It's also undesirable for a couple of reasons -
sysctls are not the appropriate interface for conveying this
information, and it causes Ubuntu to have modifications to the
x86 boot protocol which are not upstream. Therefore we should no
longer carry these changes.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>